browserify vs closure compiler
